Extract types into separate folder, add serde implementation
