+++ macros/Cargo.toml
@@ -12,6 +12,7 @@ proc-macro-error = { version = "1.0" }
 syn = { version = "2.0.114", features = ["full"] }
 quote = { version = "1" }
 bincode = { version = "2.0.1" }
+serde = { version = "1.0.228" }
 
 
 simplicityhl = { path = "../" }
